A different way to organize the code lookup table (#2)
* A different way to organize the code lookup table

* Rename

* Row org

* Extract layouts

 - use import name rather than aliases

* Simplify Layouts

* Reset keymap before switching layout

* move getQMKCode into vuex getter

* Encapsulate knowledge of internals

 - remove dependency on knowing the internals of store

* reduce code

* Better naming
