+++ src/store/modules/tester/index.js
@@ -1,0 +1,105 @@
+import defaults from '../config';
+import keys from 'lodash/keys';
+import Vue from 'vue';
+import qmkToPos from './qmk';
+import codeToPos from './codeToPos';
+import layouts from './layouts';
+import isUndefined from 'lodash/isUndefined';
+
+function reduceCodeToPos(arr) {
+  // Create look up table for Browser Code to Layout position
+  return arr.reduce((acc, code, idx) => {
+    acc[code] = idx;
+    return acc;
+  }, {});
+}
+
+const codeToPosition = {
+  ANSI: reduceCodeToPos(codeToPos.ANSI),
+  ISO: reduceCodeToPos(codeToPos.ISO)
+};
+const state = {
+  defaults,
+  codeToPosition,
+  config: Object.assign({}, defaults),
+  layout: 'ISO',
+  keymap: {},
+  layouts: {
+    ISO: layouts.ISO,
+    ANSI: layouts.ANSI
+  }
+};
+
+const getters = {
+  availableLayouts(state) {
+    return keys(state.layouts);
+  },
+  getQMKCode(state) {
+    return pos => {
+      if (isUndefined(pos)) {
+        return '';
+      }
+      return state.keymap[state.layout][pos].code;
+    };
+  },
+  activeKeymap(state) {
+    return state.keymap[state.layout];
+  },
+  activeLayoutMeta(state) {
+    return state.layouts[state.layout];
+  },
+  codeToPosition(state) {
+    return state.codeToPosition[state.layout];
+  }
+};
+
+function mapKeymap(store, arr) {
+  // Create look up table for QMK Code to Layout position
+  return arr.map(code => {
+    const meta = store.getters['keycodes/lookupKeycode'](code);
+    return {
+      ...meta
+    };
+  });
+}
+
+const actions = {
+  init({ state, commit }) {
+    const store = this;
+    commit('setKeymap', {
+      ANSI: mapKeymap(store, qmkToPos.ANSI),
+      ISO: mapKeymap(store, qmkToPos.ISO)
+    });
+    return state.keymap;
+  }
+};
+
+const mutations = {
+  setLayout(state, layout) {
+    state.layout = layout;
+  },
+  setKeymap(state, keymap) {
+    state.keymap = keymap;
+  },
+  setActive(state, { pos }) {
+    Vue.set(state.keymap[state.layout][pos], 'active', true);
+    Vue.set(state.keymap[state.layout][pos], 'detected', false);
+  },
+  setDetected(state, { pos }) {
+    Vue.set(state.keymap[state.layout][pos], 'active', false);
+    Vue.set(state.keymap[state.layout][pos], 'detected', true);
+  },
+  reset(state) {
+    state.keymap[state.layout].forEach((v, idx) => {
+      Vue.set(state.keymap[state.layout][idx], 'detected', false);
+    });
+  }
+};
+
+export default {
+  namespaced: true,
+  state,
+  getters,
+  actions,
+  mutations
+};
